StudentCareerIntelligence_BARGAW_M/requirements.txt

streamlit
pandas
numpy
scikit-learn
nltk
pdfplumber
python-docx
